requirements.txt
appdirs==1.4.4
asgiref==3.4.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.4
django-allauth==0.44.0
django-crispy-forms==1.12.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
dynaconf==3.1.4
filelock==3.0.12
hvac==0.10.14
idna==3.2
knox==0.1.14
loguru==0.5.3
oauthlib==3.1.1
pipupgrade==1.9.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
validators==0.18.2
virtualenv==20.4.7
win32-setctime==1.0.3
